--- ConsoleApp1/ConsoleApp1.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\function-interpolation\function-interpolation.csproj" />
-    <ProjectReference Include="..\linear-system\linear-system.csproj" />
-  </ItemGroup>
-
-</Project>
