+++ lab2/MainForm.resx
