+++ lab2/Properties/AssemblyInfo.cs
