+++ lab2/Properties/Resources.resx
