+++ lab3/MainWindow.xaml
@@ -1,30 +1,63 @@
-﻿<Window x:Class="lab4.MainWindow"
+﻿<Window x:Class="lab3.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:lab4"
+        xmlns:local="clr-namespace:lab3"
         mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800"
-        FontSize="15">
+        Title="MainWindow" Height="450" Width="800">
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="auto"/>
-            <RowDefinition/>
-            <RowDefinition/>
-            <RowDefinition/>
+            <RowDefinition Height="auto"/>
+            <RowDefinition Height="auto"/>
         </Grid.RowDefinitions>
 
-        <StackPanel Orientation="Horizontal" Margin="20" HorizontalAlignment="Center">
-            <TextBlock Text="n = " Height="30" VerticalAlignment="Center"/>
-            <TextBox x:Name="textBoxN" Width="30" Height="30" Margin="3" VerticalAlignment="Bottom" Text="4" TextChanged="textBoxN_TextChanged"/>
-            <TextBlock Text="f(x) = exp(-0.15x)sin(Pi(x + x^2))" Height="30"/>
+        <StackPanel Grid.Row="0" Orientation="Horizontal" Margin="20">
+            <TextBlock Text="Размерность матрицы" Height="30"/>
+            <TextBox x:Name="MatrixSizeTextBox" TextChanged="MatrixSizeTextBox_TextChanged" Width="50" Height="30" 
+                     Text="4" TextAlignment="Right"  Margin="3"/>
         </StackPanel>
-        <StackPanel Grid.Row="1" x:Name="stackPanelNodes"/>
-        <Button Grid.Row="2" x:Name="buttonStart" Content="Получить полином" Click="buttonStart_Click" Width="200" Height="30"/>
-        <StackPanel Orientation="Horizontal" Grid.Row="3" HorizontalAlignment="Center">
-            <TextBlock Text="L(x) = " Width="40" Height="30"/>
-            <TextBox x:Name="textBoxResult" IsReadOnly="True" Height="30" MaxWidth="400"/>
+
+        <StackPanel Grid.Row="1" Orientation="Horizontal" Margin="20">
+            <StackPanel Orientation="Vertical" Margin="10">
+                <TextBlock Text="Матрица A:"/>
+                <StackPanel x:Name="Matrix" Orientation="Vertical" Margin="10"/>
+            </StackPanel>
+            <StackPanel Orientation="Vertical" Margin="10">
+                <TextBlock Text="Вектор свободных коэффициентов:"/>
+                <StackPanel x:Name="FreeCoefficientsVector" Orientation="Vertical" Margin="10"/>
+            </StackPanel>
+            <StackPanel Orientation="Vertical" Margin="10">
+                <TextBlock Text="Вектор неизвестных:"/>
+                <StackPanel x:Name="ResultVector" Orientation="Vertical" Margin="10"/>
+            </StackPanel>
         </StackPanel>
-       </Grid>
+
+        <Grid Grid.Row="2" HorizontalAlignment="Stretch">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="auto"/>
+                <ColumnDefinition Width="auto"/>
+            </Grid.ColumnDefinitions>
+            <StackPanel Orientation="Horizontal" Grid.Column="0" HorizontalAlignment="Left" Margin="20">
+                <Button x:Name="MethodGaussButton" Content="Метод Гаусса" 
+                    Width="100" Height="30" Click="GaussButton_Click" 
+                    Margin="10"/>
+                <Button x:Name="IterationMethodButton" Content="Метод итераций" 
+                    Width="100" Height="30"  
+                     Margin="10" Click="IterationMethodButton_Click"/>
+            </StackPanel>
+                
+            <StackPanel x:Name="GaussResultsStackPanel" Orientation="Horizontal" Grid.Column="1" HorizontalAlignment="Right" Margin="20">
+                <StackPanel Orientation="Horizontal" Margin="10">
+                    <TextBlock Text="Определитель матрицы" Height="30"/>
+                    <TextBox x:Name="DeterminantTextBox" Width="50" Height="30" TextAlignment="Right" IsReadOnly="True" Margin="3"/>
+                </StackPanel>
+                <StackPanel Orientation="Horizontal" Margin="10">
+                    <TextBlock Text="Число обусловленности матрицы" Height="30"/>
+                    <TextBox x:Name="ConditionNumberTextBox" Width="50" Height="30" TextAlignment="Right" IsReadOnly="True" Margin="3"/>
+                </StackPanel>
+            </StackPanel>
+        </Grid>
+    </Grid>
 </Window>
