+++ lab4/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="lab3.App"
+﻿<Application x:Class="lab4.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:lab3"
+             xmlns:local="clr-namespace:lab4"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
