+++ lab4/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace lab3
+namespace lab4
 {
   /// <summary>
   /// Interaction logic for App.xaml
