+++ lab4/GraphicsWindow.xaml
