+++ lab4/GraphicsWindow.xaml.cs
