+++ lab4/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "lab2": {
+    "lab3": {
       "commandName": "Project"
     }
   }
